начал разрабатывать ТЗ-10. Где-то 2/3 - вроде бы доделал ТЗ-10, сейчас будет отладка
