+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -17,9 +17,6 @@ import java.util.stream.Collectors;
 @RestController
 @Slf4j
 public class UserController {
-    private Map<Integer, User> users = new HashMap<>();
-    private int id = 0;
-
     private final UserService userService;
 
     @Autowired
@@ -31,41 +28,15 @@ public class UserController {
     public User create(@RequestBody User user) {
         log.info("Получен POST-запрос на создание пользователя - {}", user);
 
-        try {
-            validate(user);
-        } catch (ValidationException e) {
-            log.error(e.getMessage());
-            throw new ValidationException(e.getMessage());
-        }
-
-        id++;
-        user.setId(id);
-        users.put(user.getId(), user);
-
-        return user;
+        return userService.create(user);
     }
 
     @PutMapping("/users")
     public User update(@RequestBody User user) {
         log.info("Получен PUT-запрос на обновление пользователя - {}", user);
 
-        try {
-            validate(user);
-        } catch (ValidationException e) {
-            log.error(e.getMessage());
-            throw new ValidationException(e.getMessage());
-        }
-
-        if (!users.containsKey(user.getId())) {
-            String text = "Не найден пользователь с id = " + user.getId();
-            log.error(text);
-            throw new ValidationException(text);
-        }
-        users.put(user.getId(), user);
-
-        return user;
+        return userService.update(user);
     }
-
 
     @GetMapping("/users/{id}")
     public User findById(@PathVariable("id") int userId) {
@@ -84,7 +55,7 @@ public class UserController {
 
     @GetMapping("/users")
     public List<User> getUsers() {
-        return users.values().stream().collect(Collectors.toList());
+        return userService.getAllUsers();
     }
 
     @PutMapping("/users/{id}/friends/{friendId}")
@@ -95,26 +66,5 @@ public class UserController {
     @DeleteMapping("/users/{id}/friends/{friendId}")
     public void deleteFriend(@PathVariable("id") int userId, @PathVariable("friendId") int friendId) {
         userService.removeFriend(userId, friendId);
-    }
-
-    private void validate(User user) {
-        if (user.getEmail() == null || user.getEmail().isBlank()) {
-            throw new ValidationException("Email не может быть пустым");
-        }
-        if (!user.getEmail().contains("@")) {
-            throw new ValidationException("Email должен содержать @");
-        }
-        if (user.getLogin() == null || user.getLogin().isEmpty()) {
-            throw new ValidationException("Логин не может быть пустым");
-        }
-        if (user.getLogin().contains(" ")) {
-            throw new ValidationException("Логин не может содержать пробелы");
-        }
-        if (user.getName() == null || user.getName().isEmpty()) {
-            user.setName(user.getLogin());
-        }
-        if (user.getBirthday() != null && user.getBirthday().isAfter(LocalDate.now())) {
-            throw new ValidationException("Дата рождения не может быть в будущем");
-        }
     }
 }
