+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,22 +1,57 @@
 package ru.yandex.practicum.filmorate.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
+@Slf4j
 public class UserService {
     private final UserStorage userStorage;
 
     @Autowired
     public UserService(UserStorage userStorage) {
         this.userStorage = userStorage;
+    }
+
+    public User create(User user) {
+        try {
+            validate(user);
+        } catch (ValidationException e) {
+            log.error(e.getMessage());
+            throw new ValidationException(e.getMessage());
+        }
+
+        return userStorage.create(user);
+    }
+
+    public User update(User user) {
+        try {
+            validate(user);
+        } catch (ValidationException e) {
+            log.error(e.getMessage());
+            throw new ValidationException(e.getMessage());
+        }
+
+        if (userStorage.getUserById(user.getId()) == null) {
+            String text = "Не найден пользователь с id = " + user.getId();
+            log.error(text);
+            throw new ValidationException(text);
+        }
+        return userStorage.update(user);
+    }
+
+    public List<User> getAllUsers() {
+        return userStorage.getUsersList();
     }
 
     public User getUserById(int id) {
@@ -91,5 +126,26 @@ public class UserService {
 
     private List<User> getUserListFromIdList(List<Integer> listInt) {
         return listInt.stream().map(p -> userStorage.getUserById(p)).collect(Collectors.toList());
+    }
+
+    private void validate(User user) {
+        if (user.getEmail() == null || user.getEmail().isBlank()) {
+            throw new ValidationException("Email не может быть пустым");
+        }
+        if (!user.getEmail().contains("@")) {
+            throw new ValidationException("Email должен содержать @");
+        }
+        if (user.getLogin() == null || user.getLogin().isEmpty()) {
+            throw new ValidationException("Логин не может быть пустым");
+        }
+        if (user.getLogin().contains(" ")) {
+            throw new ValidationException("Логин не может содержать пробелы");
+        }
+        if (user.getName() == null || user.getName().isEmpty()) {
+            user.setName(user.getLogin());
+        }
+        if (user.getBirthday() != null && user.getBirthday().isAfter(LocalDate.now())) {
+            throw new ValidationException("Дата рождения не может быть в будущем");
+        }
     }
 }
